LayoutTests/css3/flexbox/flexitem-expected.txt


PASS .flexbox 1 
PASS .flexbox 2 
PASS .flexbox 3 
PASS .flexbox 4 
PASS .flexbox 5 
PASS .flexbox 6 
PASS .flexbox 7 
PASS .flexbox 8 
PASS .flexbox 9 
PASS .flexbox 10 
PASS .flexbox 11 
FAIL .flexbox 12 assert_equals: 
<div class="flexbox">
  <!-- FIXME: This table should flex. -->
  <div data-expected-display="table" data-expected-width="600" style="display: inline-table"></div>
</div>
width expected 600 but got 0
PASS .flexbox 13 
PASS .flexbox 14 
PASS .flexbox 15 
PASS .flexbox 16 
PASS .flexbox 17 
PASS .flexbox 18 
button

object




button



button

object





Some text
Some more text
